backend/requirements.txt
numpy>=1.26
pillow>=10.3
pytesseract>=0.3.10
pdf2image>=1.17.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
sentence-transformers>=2.7.0
torch>=2.2,<3.0           # CPU is fine; CUDA not required
langchain>=0.2.10
faiss-cpu>=1.8.0